added some code for creating an attr macro to define days
the idea is to use a macro to define a
day (e.g. `#[aoc::day(year="2023",day="4")]`), then using that
information create a mainloop that (at compile time) uses a list of
those days and runs them depending on runtime arguments.
the code added is gotten from
[here](https://github.com/fortanix/rust-sgx/blob/jb/sgx-detect/sgxs-tools/src/sgx-detect/proc_macro.rs#L122-L294)
and will be used as a reference.