fix(clipboard): workaround for copying in linux/wayland
In wayland the supplier of clipboard data should live the whole length for when the clipboard content is needed. `wl-copy` does this under the hood by keeping something alive. However our AoC dies after it is done, so the clipboard content is gone.
To fix we make use of the wl-copy command, expecting it to exist (typical works on my machine, but that is realistically the only location it will run).